print bokehjs test output on error
